+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## TODO:
 
+## Release on 2024-10-11
+- Add thead and tbody to the toHtml() method. 
+
 ## Release on 2024-09-23
 - Improved toHtml() method.  Now allows for adding a class name, caption and 
   customized header rows.